src/requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
dj-database-url==0.5.0
Django
Django-Admin-Object-Actions
django-cors-headers==3.5.0
django-crum
django-filter
django-model-utils
graphene
graphene-django
graphql-core<3,>=2.1.0
graphql-relay<3,>=2
gunicorn==20.0.4
Pillow==7.2.0
promise==2.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
whitenoise==5.2.0
